requirements.txt
annotated-types==0.5.0
anyio==3.7.1
axiom-py==0.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dacite==1.8.1
dnspython==2.4.1
email-validator==2.0.0.post2
fastapi==0.101.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
minio==7.1.15
msgpack==1.0.5
ndjson==0.3.1
orjson==3.9.4
proto-plus==1.22.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyhumps==1.6.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.9.1
rfc3339==6.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
